requirements.txt
Django==3.1.7
psycopg2==2.8.6
djangorestframework
django-filter==2.4.0